Move state accessors into their own file
