+++ packages/perseus/src/widgets/interactive-graphs/reducer/initialize-graph-state.test.ts
@@ -1,7 +1,6 @@
 import invariant from "tiny-invariant";
 
 import {
-    getGradableGraph,
     initializeGraphState,
 } from "./initialize-graph-state";
 
@@ -405,33 +404,5 @@ describe("initializeGraphState for sinusoid graphs", () => {
             [0, 0],
             [3, 2],
         ]);
-    });
-});
-
-// FIXME: move this
-describe("getGradableGraph", () => {
-    /**
-     * Originally `getGradableGraph` was returning a PerseusGraphType with just a
-     * `type` property, stripping everything else off of `initialGraph`.
-     * This caused the editor to keep resetting certain properties (ie `numSegments`).
-     */
-    it("regression LEMS-1935: false hasBeenInteractedWith returns full initial graph", () => {
-        const state: InteractiveGraphState = {
-            type: "segment",
-            coords: [],
-            hasBeenInteractedWith: false,
-            range: [
-                [-10, 10],
-                [-10, 10],
-            ],
-            snapStep: [1, 1],
-        };
-        const initialGraph: PerseusGraphType = {
-            type: "segment",
-            numSegments: 4,
-        };
-        const result = getGradableGraph(state, initialGraph);
-
-        expect(result).toEqual(initialGraph);
     });
 });
