+++ packages/perseus/src/widgets/interactive-graphs/reducer/initialize-graph-state.ts
@@ -181,88 +181,6 @@ const getDefaultPoints = ({
     return normalizePoints(range, step, newCoords);
 };
 
-// TS v4 doesn't narrow return types, while v5 does.
-// Instead of updating to v5, using generic type to relate input and output types.
-export function getGradableGraph(
-    state: InteractiveGraphState,
-    initialGraph: PerseusGraphType,
-): PerseusGraphType {
-    if (!state.hasBeenInteractedWith) {
-        return {...initialGraph};
-    }
-
-    if (
-        initialGraph.type === "linear-system" &&
-        state.type === "linear-system"
-    ) {
-        return {
-            ...initialGraph,
-            coords: state.coords,
-        };
-    }
-
-    if (state.type === "segment" && initialGraph.type === "segment") {
-        return {
-            ...initialGraph,
-            coords: state.coords,
-        };
-    }
-
-    if (state.type === "linear" && initialGraph.type === "linear") {
-        return {
-            ...initialGraph,
-            coords: state.coords,
-        };
-    }
-
-    if (state.type === "ray" && initialGraph.type === "ray") {
-        return {
-            ...initialGraph,
-            coords: state.coords,
-        };
-    }
-
-    if (state.type === "polygon" && initialGraph.type === "polygon") {
-        return {
-            ...initialGraph,
-            coords: state.coords,
-        };
-    }
-
-    if (state.type === "point" && initialGraph.type === "point") {
-        return {
-            ...initialGraph,
-            coords: state.coords,
-        };
-    }
-
-    if (state.type === "circle" && initialGraph.type === "circle") {
-        return {
-            ...initialGraph,
-            center: state.center,
-            radius: getRadius(state),
-        };
-    }
-
-    if (state.type === "quadratic" && initialGraph.type === "quadratic") {
-        return {
-            ...initialGraph,
-            coords: state.coords,
-        };
-    }
-
-    if (state.type === "sinusoid" && initialGraph.type === "sinusoid") {
-        return {
-            ...initialGraph,
-            coords: state.coords,
-        };
-    }
-
-    throw new Error(
-        "Mafs is not yet implemented for graph type: " + initialGraph.type,
-    );
-}
-
 type getDefaultSegmentsArg = {
     graph: PerseusGraphTypeSegment;
     range: InitializeGraphStateParam["range"];
@@ -458,18 +376,3 @@ const getCircleCoords = (
         radiusPoint: [1, 0],
     };
 };
-
-// FIXME move this
-/**
- * determine radius of a circle graph
- *
- * @param graph - the graph object containing the circle
- * @returns the radius of the circle
- */
-export function getRadius(graph: CircleGraphState): number {
-    const [centerX, centerY] = graph.center;
-    const [edgeX, edgeY] = graph.radiusPoint;
-    return Math.sqrt(
-        Math.pow(edgeX - centerX, 2) + Math.pow(edgeY - centerY, 2),
-    );
-}
